refactor: split intent-agent type definitions
